requirements.txt
fastapi
uvicorn
pdfminer.six
python-docx
openai
pinecone
langchain
tiktoken
python-multipart
sentence_transformers